third_party/blink/web_tests/editing/style/apply-font-size-to-multiple-nodes-expected.txt

Test that WebKit does not add multiple span or font elements when applying font-size to multiple nodes.

You should see exactly one font element with size="7" and exactly one span:
| <font>
|   size="7"
|   "
<#selection-anchor>A man with 2
"
|   <span>
|     style="font-family: Courier New, Courier"
|     "font faces"
|   <br>
|   "
Death comes in all
"
|   "font sizes<#selection-focus>"
| "
"
